fix: consolidate julia env handling
