+++ scripts/terminalserver/terminalserver.jl
@@ -1,48 +1,9 @@
-@static if VERSION < v"1.8.0"
-    import Pkg
-end
-
-# this script basially only handles `ARGS`
-let distributed = Base.PkgId(Base.UUID("8ba89e20-285c-5b6f-9357-94700520ee1b"), "Distributed")
-    version_specific_env_path = joinpath(@__DIR__, "..", "environments", "terminalserver", "v$(VERSION.major).$(VERSION.minor)")
-    if !isdir(version_specific_env_path)
-        version_specific_env_path = joinpath(@__DIR__, "..", "environments", "terminalserver", "fallback")
-    end
-
-    activate_env = () -> begin
-        @static if VERSION < v"1.8.0"
-            Pkg.activate(version_specific_env_path)
-        else
-            Base.set_active_project(joinpath(version_specific_env_path, "Project.toml"))
-        end
-    end
-    deactivate_env = () -> begin
-        @static if VERSION < v"1.8.0"
-            # TODO Maybe surpress output if all goes well
-            Pkg.activate(project_path)
-        else
-            Base.set_active_project(joinpath(project_path, "Project.toml"))
-        end
-    end
+# this script loads VSCodeServer and handles ARGS
+let
+    include("load_vscodeserver.jl")
 
     args = [popfirst!(Base.ARGS) for _ in 1:9]
     conn_pipename, debug_pipename, telemetry_pipename, project_path = args[1:4]
-
-    activate_env()
-
-    if haskey(Base.loaded_modules, distributed) && (Distributed = Base.loaded_modules[distributed]).nprocs() > 1
-        # TODO Move this logic over to the project loading mechanism
-        Distributed.remotecall_eval(Main, 1:Distributed.nprocs(), :($(activate_env)()))
-        try
-            using VSCodeServer
-        finally
-            Distributed.remotecall_eval(Main, 1:Distributed.nprocs(), :($(deactivate_env)()))
-        end
-    else
-        using VSCodeServer
-    end
-
-    deactivate_env()
 
     # load Revise ?
     if "USE_REVISE=true" in args
